BUG: Ensure part_prescence matches shape id
When drawing the shapes with store_and_draw, use the shape id we
have for in shapes_symb_output as the part_number to make sure they
correspond to each other. This makes sure our use of extract_part
in check_bordering and check_containing will extract the correct
shape for us.

Haven't seen a problem arise from this bug, since the part_number
we were supplying was in all cases the same as the shape id, but
this fix will make sure it can't occur in the future either.