BUG: Fix initialisation in extract_part and any_content_collides
We were not initialising our loop variable to 0 and were using
whatever its previous value in memory was. This will have meant
only a subset of pixels in the vignette will have been checked,
which will have lead to incorrect parsing output.

This was not spotted previously since most of the time memory
begins with the value 0.